@@ -8,31 +8,31 @@
     "@rails/activestorage": "^6.0.0",
     "@rails/ujs": "^6.0.0",
     "@rails/webpacker": "4.3.0",
-    "activeadmin_addons": "^1.0.0",
     "@tailwindcss/postcss7-compat": "^2.0.4",
+    "activeadmin_addons": "^1.0.0",
     "arctic_admin": "^4.0.1",
     "autoprefixer": "^9",
     "postcss": "^7",
     "tailwindcss": "npm:@tailwindcss/postcss7-compat",
     "turbolinks": "^5.2.0",
-    "vue": "^2.6.14",
-    "vue-loader": "^15.9.7",
-    "vue-template-compiler": "^2.6.14"
+    "vue": "^3.1.5",
+    "vue-loader": "^16.4.1"
   },
   "version": "0.1.0",
   "devDependencies": {
-    "@vue/test-utils": "^1.2.2",
+    "@vue/compiler-sfc": "^3.1.5",
+    "@vue/test-utils": "^2.0.0-rc.11",
     "autoprefixer": "^9",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-jest": "^27.0.6",
+    "babel-jest": "^26.5.6",
     "eslint": "^7.32.0",
     "eslint-plugin-import": "^2.23.4",
     "eslint-plugin-vue": "^7.15.1",
-    "jest": "^27.0.6",
+    "jest": "^26.5.6",
     "jest-serializer-vue": "^2.0.2",
     "postcss": "^7",
     "stylelint": "^13.13.1",
-    "vue-jest": "^3.0.7",
+    "vue-jest": "^5.0.0-alpha.6",
     "webpack-dev-server": "^3.11.2"
   },
   "engines": {
